More work on session 2, some updates to common_predoc.  Small updates to code

+++ code/1_control_structure/lectures/factorial.cpp
@@ -1,6 +1,6 @@
 #include <iostream>
-
-unsigned long factorial(const unsigned long n)
+// Notice we're using double here as unsigned int could be too small
+double factorial(const unsigned int n)
 {
   if(n > 1)
     return (n * factorial(n - 1));
